+++ walk_search_history/state.py
@@ -6,21 +6,26 @@ __copyright__ = '2011, Grant Drake'
 # calibre Python 3 compatibility.
 from six import text_type as unicode
 
+MAX_HISTORY = 25
+
+def get_initial_search_items(search):
+    # Seed our history list with the contents of the search combobox history as a starting point
+    snapshot = list()
+    if search:
+        for i in range(search.count()):
+            snapshot.append(unicode(search.itemText(i)))
+    return snapshot
+
+
 class SearchHistoryState:
-    # Keep a history of searches.
-    # This class is our "visited" history, and is used for the dropdown menu
-    # It is also the base class for our NavigationSearchHistoryState class which is
-    # used for the fwd/bwd navigation.
-
-    MAX_HISTORY = 25
-
+    '''
+    Keep a history of searches.
+    This class is our "visited" history, and is used for the dropdown menu
+    This should always look the same as the contents of the calibre search dropdown
+    other than our specified threshold of 25 items.
+    '''
     def __init__(self, search):
-        self.clear()
-        self.snapshot = list()
-        # Seed our history list with the contents of the search combobox history as a starting point
-        if search:
-            for i in range(search.count()):
-                self.snapshot.append(unicode(search.itemText(i)))
+        self.snapshot = get_initial_search_items(search)
 
     def clear(self):
         self.snapshot = []
@@ -28,50 +33,63 @@ class SearchHistoryState:
     def items(self):
         return self.snapshot
 
-    def insert(self, text):
+    def append(self, text):
         # If the item exists in our history, move it to the top
         # Otherwise add it at the top
         if text in self.snapshot:
             self.snapshot.remove(text)
         self.snapshot.insert(0, text)
         # Make sure we don't keep too much history...
-        while len(self.snapshot) > self.MAX_HISTORY:
+        while len(self.snapshot) > MAX_HISTORY:
             self.snapshot.pop()
 
 
-class NavigationSearchHistoryState(SearchHistoryState):
-    # Keep track of the recent searches allowing fwd/bwd navigation.
-    # When the user issues a new search (breaking the sequence) all
-    # "forward" entries if any will be removed.
+class NavigationSearchHistoryState:
+    '''
+    Keep track of the recent searches allowing fwd/bwd navigation.
+    Starts off the same as the calibre search history at plugin
+    startup then diverges over time based on user navigation in
+    a similar way to how a web browser navigation works.
+    When the user issues a new search (breaking the sequence) all
+    "forward" entries if any will be removed.
+    '''
 
     def __init__(self, search):
-        SearchHistoryState.__init__(self, search)
-        # Ensure we have a blank search (as this is system startup) to move forward to if go back
-        self.snapshot.insert(0, '')
-        self.position = 0
+        self.position = -1
+        self.snapshot = get_initial_search_items(search)
 
     def clear(self):
-        SearchHistoryState.clear(self)
+        self.position = -1
+        self.snapshot = []
+
+    def reset_after_empty_search(self):
+        # We don't want to insert an empty search in the history list to navigate to.
+        # However we do want to make sure that if the user navigates backwards after a
+        # empty search that they hit the topmost item in the search list.
+        # My decision is that a user applying empty text search should remove any
+        # history "forward" of the current position.
+        if self.position > 0:
+            for _i in range(self.position):
+                self.snapshot.pop(0)
+        # Now ensure that if they navigate "back" they get the new topmost item.
         self.position = -1
 
-    def insert(self, text):
+    def append(self, text):
         # Override the base behaviour slightly
         # If we have navigated backwards, remove searches that are considered in future
         if self.position > 0:
-            for i in range(self.position):
+            for _i in range(self.position):
                 self.snapshot.pop(0)
         # It is ok to have duplicates in our list as this reflects order of searches
         self.snapshot.insert(0, text)
         # Make sure we don't keep too much history...
-        while len(self.snapshot) > self.MAX_HISTORY:
+        while len(self.snapshot) > MAX_HISTORY:
             self.snapshot.pop()
-        # Reset our search position
+        # Reset our search position to reflect the current text
         self.position = 0
 
     def get_current(self):
-        print('get_current:',self.snapshot)
         if self.position != -1:
-            print('get_current at position:',self.snapshot[self.position])
             return self.snapshot[self.position]
         return ''